+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file. This projec
 
 ## [Unreleased]
 
+### Added
+
+- Integrated our block plugin support package into the plugin base for easier use and maintenance.
+
 ### Fixed
 
 - AssetManager was not hooking the static implementation but the abstract.